update discussion model
changed discussions to discussion
